requirements.txt
yfinance>=0.2.43
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.11.0
requests>=2.32.3
python-dotenv>=1.0.1
ta>=0.11.0
feedparser>=6.0.12
schedule>=1.2.2
beautifulsoup4>=4.13.5
pytz>=2025.2
ccxt>=4.1.0
websocket-client>=1.6.0
flask>=3.0.0
discord.py>=2.3.0
vaderSentiment>=3.3.2
textblob>=0.17.1
